pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Schislyaev <pschhhh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.5"
aio-pika = "^9.0.4"
aioredis = "^2.0.1"
sentry-sdk = "^1.16.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
motor = "^3.1.1"
mongoengine = "^0.27.0"
beanie = "^1.17.0"
email-validator = "^1.3.1"
pyjwt = "^2.6.0"
grpcio-tools = "^1.51.3"
orjson = "^3.8.7"
apscheduler = "^3.10.1"
jinja2 = "^3.1.2"
croniter = "^1.3.8"
types-croniter = "^1.3.2.6"
websockets = "^10.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.0.1"
flake8-html = "^0.4.3"
types-croniter = "^1.3.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
